Change DeltaTime implementation

+++ src/Engine/Application/Application.h
@@ -44,7 +44,7 @@ class Application {
    protected:
     std::string m_ProjectName;
     std::unordered_map<std::string, std::vector<GameObject*>> m_Rooms;
-    float m_LastTick;
+    Uint32 m_LastTick;
     bool m_is_paused;
     bool m_has_focus;
     int m_Frame;
